+++ frontend/src/components/Wifi/WifiHostapd.js
@@ -21,11 +21,17 @@ import {
   Button,
   ButtonIcon,
   ButtonText,
+  Checkbox,
+  CheckboxIcon,
+  CheckboxIndicator,
+  CheckboxLabel,
   FormControl,
   FormControlLabel,
   FormControlLabelText,
   FormControlError,
   FormControlErrorText,
+  FormControlHelper,
+  FormControlHelperText,
   Heading,
   HStack,
   Input,
@@ -121,7 +127,10 @@ const WifiHostapd = (props) => {
     'ieee80211ax',
     'he_su_beamformer',
     'he_su_beamformee',
-    'he_mu_beamformer'
+    'he_mu_beamformer',
+    'rssi_reject_assoc_rssi',
+    'rssi_reject_assoc_timeout',
+    'rssi_ignore_probe_request'
   ]
   const canEdit = canEditInt.concat(canEditString)
 
@@ -241,6 +250,16 @@ const WifiHostapd = (props) => {
     wifiAPI
       .config(iface)
       .then((conf) => {
+        // Initialize RSSI fields if they don't exist
+        if (conf.rssi_reject_assoc_rssi === undefined) {
+          conf.rssi_reject_assoc_rssi = 0
+        }
+        if (conf.rssi_reject_assoc_timeout === undefined) {
+          conf.rssi_reject_assoc_timeout = 30
+        }
+        if (conf.rssi_ignore_probe_request === undefined) {
+          conf.rssi_ignore_probe_request = 0
+        }
         setConfig(sortConf(conf))
       })
       .catch((err) => {
@@ -279,6 +298,17 @@ const WifiHostapd = (props) => {
       He_mu_beamformer: parseInt(inconfig.he_mu_beamformer)
     }
 
+    // Add RSSI fields if they exist
+    if (inconfig.rssi_reject_assoc_rssi !== undefined) {
+      data.Rssi_reject_assoc_rssi = parseInt(inconfig.rssi_reject_assoc_rssi)
+    }
+    if (inconfig.rssi_reject_assoc_timeout !== undefined) {
+      data.Rssi_reject_assoc_timeout = parseInt(inconfig.rssi_reject_assoc_timeout)
+    }
+    if (inconfig.rssi_ignore_probe_request !== undefined) {
+      data.Rssi_ignore_probe_request = parseInt(inconfig.rssi_ignore_probe_request)
+    }
+
     wifiAPI.updateConfig(iface, data).then((curConfig) => {
       setConfig(sortConf(curConfig))
     })
@@ -296,6 +326,17 @@ const WifiHostapd = (props) => {
       He_su_beamformer: parseInt(config.he_su_beamformer),
       He_su_beamformee: parseInt(config.he_su_beamformee),
       He_mu_beamformer: parseInt(config.he_mu_beamformer)
+    }
+
+    // Add RSSI fields if they exist
+    if (config.rssi_reject_assoc_rssi !== undefined) {
+      data.Rssi_reject_assoc_rssi = parseInt(config.rssi_reject_assoc_rssi)
+    }
+    if (config.rssi_reject_assoc_timeout !== undefined) {
+      data.Rssi_reject_assoc_timeout = parseInt(config.rssi_reject_assoc_timeout)
+    }
+    if (config.rssi_ignore_probe_request !== undefined) {
+      data.Rssi_ignore_probe_request = parseInt(config.rssi_ignore_probe_request)
     }
 
     wifiAPI.updateConfig(iface, data).then((curConfig) => {