+++ installer/pi-download-image.sh
@@ -5,12 +5,13 @@ mkdir data
 # 25.04 is not yet ready due to mt7915e, r8125 driver bugs
 # 24.10 is needed for cm5 support
 VERSION="24.10"
-IMG="ubuntu-${VERSION}-preinstalled-server-arm64+raspi.img.xz"
+IMG="ubuntu-${VERSION}-beta-preinstalled-server-arm64+raspi.img.xz"
 
 cd ./data
 
 if [ ! -f $IMG ]; then
-  wget -q "https://cdimage.ubuntu.com/releases/${VERSION}/release/${IMG}"
+  #temp until 25.04 fixes drivers for r8125, mt7915
+  wget -q "https://old-releases.ubuntu.com/releases/oracular/${IMG}"
 fi
 
 xzcat $IMG > spr.clean.img