+++ RELEASE-NOTES.md
@@ -1,5 +1,14 @@
 # Secure Programmable Router (SPR) Release Notes
 
+## v1.0.16
+**Improvements**
+- Added RSSI threshold support for limiting when devices connect
+- Added support for multiple TLS dns upstreams to coredns
+- Support open ssid for uplink connections
+**Fixes**
+- Smoothed out OTP flow for plugins, fixed OTP login, added OTP rate limits
+- Golang module updates
+
 ## v1.0.15
 **Fixes**
 - Bump build for guest ap fix